# internal/users/protos/openapi_types.gen.go
// Package protos provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package protos

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	// Avatar User avatar URL (optional)
	Avatar *string `json:"avatar,omitempty"`

	// Name User name
	Name string `json:"name"`
}

// CreateUserResponse defines model for CreateUserResponse.
type CreateUserResponse struct {
	User *User `json:"user,omitempty"`
}

// DeleteUserResponse defines model for DeleteUserResponse.
type DeleteUserResponse struct {
	// Message Deletion result message
	Message *string `json:"message,omitempty"`

	// Success Whether the deletion was successful
	Success *bool `json:"success,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *string `json:"code,omitempty"`

	// Message Error message
	Message *string `json:"message,omitempty"`
}

// GetUserResponse defines model for GetUserResponse.
type GetUserResponse struct {
	User *User `json:"user,omitempty"`
}

// UpdateUserRequest defines model for UpdateUserRequest.
type UpdateUserRequest struct {
	// Avatar User avatar URL (optional)
	Avatar *string `json:"avatar,omitempty"`

	// Balance User balance (optional, can be negative)
	Balance *int32 `json:"balance,omitempty"`

	// Name User name (optional)
	Name *string `json:"name,omitempty"`
}

// UpdateUserResponse defines model for UpdateUserResponse.
type UpdateUserResponse struct {
	User *User `json:"user,omitempty"`
}

// User defines model for User.
type User struct {
	// Avatar User avatar URL
	Avatar *string `json:"avatar,omitempty"`

	// Balance User balance (can be negative)
	Balance *int32 `json:"balance,omitempty"`

	// CreatedAt Creation timestamp
	CreatedAt *int64 `json:"created_at,omitempty"`

	// Id User ID
	Id *string `json:"id,omitempty"`

	// Name User name
	Name *string `json:"name,omitempty"`

	// UpdatedAt Last update timestamp
	UpdatedAt *int64 `json:"updated_at,omitempty"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = CreateUserRequest

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = UpdateUserRequest