# internal/wallet/protos/openapi_types.gen.go
// Package protos provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package protos

// Defines values for TransactionType.
const (
	DEPOSIT     TransactionType = "DEPOSIT"
	TRANSFERIN  TransactionType = "TRANSFER_IN"
	TRANSFEROUT TransactionType = "TRANSFER_OUT"
	UNKNOWN     TransactionType = "UNKNOWN"
	WITHDRAW    TransactionType = "WITHDRAW"
)

// DepositRequest defines model for DepositRequest.
type DepositRequest struct {
	// Amount Amount to deposit (must be positive)
	Amount int32 `json:"amount"`

	// Description Transaction description (optional)
	Description *string `json:"description,omitempty"`
}

// DepositResponse defines model for DepositResponse.
type DepositResponse struct {
	// NewBalance Balance after deposit
	NewBalance  *int32       `json:"new_balance,omitempty"`
	Transaction *Transaction `json:"transaction,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code *string `json:"code,omitempty"`

	// Message Error message
	Message *string `json:"message,omitempty"`
}

// GetBalanceResponse defines model for GetBalanceResponse.
type GetBalanceResponse struct {
	// Balance Current wallet balance
	Balance *int32 `json:"balance,omitempty"`

	// UserId User ID
	UserId *string `json:"user_id,omitempty"`
}

// GetTransactionsResponse defines model for GetTransactionsResponse.
type GetTransactionsResponse struct {
	// Total Total number of transactions
	Total *int32 `json:"total,omitempty"`

	// Transactions List of transactions
	Transactions *[]Transaction `json:"transactions,omitempty"`
}

// Transaction defines model for Transaction.
type Transaction struct {
	// Amount Transaction amount (positive for credit, negative for debit)
	Amount *int32 `json:"amount,omitempty"`

	// BalanceAfter Balance after transaction
	BalanceAfter *int32 `json:"balance_after,omitempty"`

	// BalanceBefore Balance before transaction
	BalanceBefore *int32 `json:"balance_before,omitempty"`

	// CreatedAt Transaction timestamp
	CreatedAt *int64 `json:"created_at,omitempty"`

	// Description Transaction description
	Description *string `json:"description,omitempty"`

	// Id Transaction ID
	Id *string `json:"id,omitempty"`

	// RelatedUserId Related user ID (for transfers)
	RelatedUserId *string `json:"related_user_id,omitempty"`

	// Type Transaction type
	Type *TransactionType `json:"type,omitempty"`

	// UserId User ID
	UserId *string `json:"user_id,omitempty"`
}

// TransactionType Transaction type
type TransactionType string

// TransferRequest defines model for TransferRequest.
type TransferRequest struct {
	// Amount Amount to transfer per user (must be positive)
	Amount int32 `json:"amount"`

	// Description Transaction description (optional)
	Description *string `json:"description,omitempty"`

	// FromUserId Sender user ID
	FromUserId string `json:"from_user_id"`

	// ToUserIds Receiver user IDs (can be one or multiple)
	ToUserIds []string `json:"to_user_ids"`
}

// TransferResponse defines model for TransferResponse.
type TransferResponse struct {
	// FromNewBalance Sender's new balance
	FromNewBalance *int32 `json:"from_new_balance,omitempty"`

	// Results Transfer results for each receiver
	Results *[]TransferResult `json:"results,omitempty"`

	// Transactions Transactions for each receiver
	Transactions *[]Transaction `json:"transactions,omitempty"`
}

// TransferResult defines model for TransferResult.
type TransferResult struct {
	// ErrorMessage Error message if failed
	ErrorMessage *string `json:"error_message,omitempty"`

	// Success Whether transfer to this user succeeded
	Success *bool `json:"success,omitempty"`

	// ToNewBalance Receiver's new balance
	ToNewBalance *int32 `json:"to_new_balance,omitempty"`

	// ToUserId Receiver user ID
	ToUserId *string `json:"to_user_id,omitempty"`
}

// WithdrawRequest defines model for WithdrawRequest.
type WithdrawRequest struct {
	// Amount Amount to withdraw (must be positive)
	Amount int32 `json:"amount"`

	// Description Transaction description (optional)
	Description *string `json:"description,omitempty"`
}

// WithdrawResponse defines model for WithdrawResponse.
type WithdrawResponse struct {
	// NewBalance Balance after withdrawal
	NewBalance  *int32       `json:"new_balance,omitempty"`
	Transaction *Transaction `json:"transaction,omitempty"`
}

// GetTransactionsParams defines parameters for GetTransactions.
type GetTransactionsParams struct {
	// Limit Maximum number of transactions to return
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Offset for pagination
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// TransferJSONRequestBody defines body for Transfer for application/json ContentType.
type TransferJSONRequestBody = TransferRequest

// DepositJSONRequestBody defines body for Deposit for application/json ContentType.
type DepositJSONRequestBody = DepositRequest

// WithdrawJSONRequestBody defines body for Withdraw for application/json ContentType.
type WithdrawJSONRequestBody = WithdrawRequest